Cleaned stuff up /a lot/.
Split everything dumb out of god-files val.h and val.c.

Each type gets it's own file (if it needs it).

val.c is now the file for ag_val stuff only.

Type definitions from val.h are now in types.h.

errors.h got subsumed into error.h, the header for error.h, which is
for val_err's.

Fixed some .deps stuff.

Made ag_sprint in print.c sane for val_list's by using
open_memstream(3) in implementation.  It's part of POSIX 2008.

Added macros to Makefile.cinclude so that *asprintf and
open_*memstream are defined.  This now requires both gnu extensions in
glibc and a very recent POSIX version that few besides Linux support.
That isn't a problem with me at the moment, though I wonder if it will
build under Cygwin.

Going to try to unify nil and lists in the reader now.

Compiles and runs to specification as is.

	new file:   .deps/error.P
	modified:   .deps/eval.P
	new file:   .deps/list.P
	new file:   .deps/print.P
	new file:   .deps/read.P
	modified:   .deps/silveretta.P
	modified:   .deps/string-helpers.P
	modified:   .deps/val.P
	modified:   Makefile
	modified:   Makefile.cincludes
	modified:   Project.ede
	new file:   error.c
	modified:   eval.c
	renamed:    include/errors.h -> include/error.h
	modified:   include/eval.h
	new file:   include/list.h
	deleted:    include/macros.h
	new file:   include/print.h
	new file:   include/read.h
	modified:   include/string-helpers.h
	new file:   include/symbol.c
	renamed:    include/val.h -> include/types.h
	modified:   include/val.h
	new file:   list.c
	new file:   print.c
	new file:   read.c
	modified:   silveretta.c
	modified:   string-helpers.c
	modified:   val.c